4/ds_services/services/intent_catcher/pyproject.toml
[tool.poetry]
name = "intent-catcher"
version = "0.1.0"
description = ""
authors = ["Denis Kuznetsov <kunetsov.den.p@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langserve = "^0.2.2"
sse-starlette = "^2.1.3"
fastapi = {extras = ["standard"], version = "^0.112.0"}
pytest = "^8.3.2"
langchain = "^0.2.13"
services-api = {path = "/common_packages/services_api", develop = true}
xeger = "^0.4.0"
pydantic = "^2.8.2"
torch = "^2.4.0"
pandas = "^2.2.2"
tqdm = "^4.66.5"
numpy = "^1.26.4"
scikit-learn = "^1.5.1"
langchain-huggingface = "^0.0.3"
faiss-cpu = "^1.8.0.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
